refactor: creation of ui utilites package
